small update

+++ more_malloc_free/0-malloc_checked.c
@@ -12,7 +12,7 @@ void *malloc_checked(unsigned int b)
 {
 	unsigned int i;
 
-	i = malloc(b);
+	i =(unsigned int)malloc(b);
 
 	if (i == NULL)
 	{
